Rename operator_estimation -> observable_estimation (#138)
