+++ forest/benchmarking/tests/test_direct_fidelity_estimation.py
@@ -5,7 +5,7 @@ from pyquil import Program
 from pyquil.api import BenchmarkConnection
 from pyquil.gates import *
 from pyquil.numpy_simulator import NumpyWavefunctionSimulator
-from forest.benchmarking.operator_estimation import _one_q_state_prep
+from forest.benchmarking.observable_estimation import _one_q_state_prep
 
 from numpy.testing import assert_almost_equal, assert_allclose
 from test_process_tomography import test_qc
