+++ forest/benchmarking/tomography.py
@@ -13,7 +13,7 @@ from pyquil.unitary_tools import lifted_pauli as pauli2matrix, lifted_state_oper
 import forest.benchmarking.distance_measures as dm
 from forest.benchmarking.utils import all_traceless_pauli_terms
 from forest.benchmarking.superoperator_tools import vec, unvec, proj_choi_to_physical
-from forest.benchmarking.operator_estimation import ExperimentSetting, ObservablesExperiment, \
+from forest.benchmarking.observable_estimation import ExperimentSetting, ObservablesExperiment, \
     ExperimentResult, SIC0, SIC1, SIC2, SIC3, plusX, minusX, plusY, minusY, plusZ, minusZ, \
     TensorProductState, zeros_state
